Fix Python release snapshot removal handling
Update Python roundup handling so that SNAPSHOT tags are, again, removed
during out build. Changes in `pds-github-util` updated the Python snapshot
release handling to use "SHAPSHOT" like our Java releases. It was missed
that the check for these tags was hard coded to "*dev*" here. We may
want to update this in the future so it's not hard coded. Or that might
be wasted worked and not matter.

This addresses one of the issues raised in #52